+++ src/main/java/com/multicampus/matchcode/model/entity/MatchDTO.java
@@ -26,5 +26,4 @@ public class MatchDTO {
     private int restrictionMinRate;
     private int restrictionMaxRate;
     private int Status;
-
 }
