+++ src/main/java/com/multicampus/matchcode/model/entity/MatchMemberDTO.java
@@ -19,6 +19,4 @@ public class MatchMemberDTO {
 
     private long matchId;
     private long userId;
-
-
 }
