+++ src/main/java/com/multicampus/matchcode/model/entity/MemberDTO.java
@@ -13,7 +13,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@Data
 public class MemberDTO {
 
     @Id
