+++ src/main/java/com/multicampus/matchcode/model/request/khj/MemberAndPointRequest.java
@@ -27,6 +27,5 @@ public class MemberAndPointRequest {
     public PointDTO getPointDTO() {
         return pointDTO;
     }
-    //막상 해놓고보니 request 클래스는 사용을 안하는데...? 필요가 없다....뭐지?
 
 }
